requirements.txt
absl-py==1.4.0
aif360==0.5.0
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
decorator==5.1.1
docker-pycreds==0.4.0
filelock==3.10.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
gitdb==4.0.10
gitpython==3.1.30
google-auth-oauthlib==0.4.6
google-auth==2.16.0
graphviz==0.20.1
grpcio==1.51.1
h5py==3.8.0
higher==0.2.1
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.0.0
ipdb==0.13.13
ipython==7.34.0
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
markdown==3.4.1
markupsafe==2.1.2
matplotlib-inline==0.1.6
matplotlib==3.5.3
multidict==6.0.4
nltk==3.8.1
numpy==1.21.6
oauthlib==3.2.2
openai-clip==1.0.1
ordered-set==4.1.0
packaging==23.0
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
pip==22.3.1
prompt-toolkit==3.0.38
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydeprecate==0.3.2
pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytorch-metric-learning==2.0.1
pytz==2022.7.1
pyyaml==6.0
regex==2022.10.31
requests-oauthlib==1.3.1
requests==2.28.2
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.14.0
setproctitle==1.3.2
setuptools==59.5.0
six==1.16.0
smmap==5.0.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.2
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torchmetrics==0.11.1
torchvision
tqdm==4.64.1
traitlets==5.9.0
transformers==4.27.2
typing-extensions==4.4.0
urllib3==1.26.14
wandb==0.13.9
wcwidth==0.2.6
werkzeug==2.2.2
wheel==0.38.4
yarl==1.8.2
zipp==3.12.0